Move src/lib/exceptions -> src/exception
